requirements.txt
-U scikit-learn
numpy
pandas
matplotlib
streamlit
streamlit>=1.25.0
pandas
boto3
langchain>=0.1.0
langchain-community
langchain-core
langchain-huggingface
sentence-transformers
faiss-cpu
python-dotenv


